break date and comment helpers into new class
