+++ app/cardrecord.js
@@ -1,16 +1,11 @@
 "use strict";
-
-var instadate = require("instadate");
 var yaml = require('js-yaml');
-var moment = require("moment");
-var fedHolidays = require('@18f/us-federal-holidays');
 var Q = require('q');
+var moment = require("moment");
 
-var d = new Date();
-var holidays = fedHolidays.allForYear(d.getFullYear());
-
+var DateCommentHelpers = require("./date-comment-helpers.js");
+var DCH = new DateCommentHelpers();
 var MyTrello = require("./my-trello.js");
-
 
 class CardRecorder extends MyTrello {
     constructor(yaml_file, board) {
@@ -51,11 +46,11 @@ class CardRecorder extends MyTrello {
           var hasMoved = false;
           var daysSinceUpdate = false;
           if(cardActions.updates.length > 0){
-            hasMoved = self.hasMovedCheck(cardActions.updates);
+            hasMoved = DCH.hasMovedCheck(cardActions.updates);
             daysSinceUpdate = now.diff(moment(cardActions.updates[0].date), 'days');
           }
-          var totDays = (cardActions.comments.length > 0)? self.calcTotalDays(cardActions.comments, now) : 0;
-          var prevMove = self.findPrevMoveDateFromComments({commentList: cardActions.comments, "actionList": cardActions.updates, "createActionDate": cardActions.created.date});
+          var totDays = (cardActions.comments.length > 0)? DCH.calcTotalDays(cardActions.comments, now) : 0;
+          var prevMove = DCH.findPrevMoveDateFromComments({commentList: cardActions.comments, "actionList": cardActions.updates, "createActionDate": cardActions.created.date});
           self.inFinalList(card.idList)
           .then(function(finalList){
             self.decideCommentType(card,
@@ -71,28 +66,17 @@ class CardRecorder extends MyTrello {
           });
       });
     return deferred.promise;
-  }
+    }
 
-  cardFilters(cardActions, actionTypeArray){
-    var actions = {};
-    actionTypeArray.forEach(function(actionType){
-      actions[actionType.name] = cardActions.filter(function(action){
-        return action.type == actionType.type;
+    cardFilters(cardActions, actionTypeArray){
+      var actions = {};
+      actionTypeArray.forEach(function(actionType){
+        actions[actionType.name] = cardActions.filter(function(action){
+          return action.type == actionType.type;
+        });
       });
-    });
-    return actions;
-
-    // var comments = cardActions.filter(function(action){
-    //   return action.type == 'commentCard';
-    // });
-    // var updateActions = cardActions.filter(function(action){
-    //   return action.type == 'updateCard';
-    // });
-    // var createAction = cardActions.filter(function(action){
-    //   return action.type =='createCard';
-    // });
-    // return {"comments": comments, "updates": updateActions, "created": createAction}
-  }
+      return actions;
+    }
 
     deleteCurrentComment(comments) {
         var deferred = Q.defer(),
@@ -117,49 +101,6 @@ class CardRecorder extends MyTrello {
         return deferred.promise;
     }
 
-    checkCommentsForDates(commentList, latest){
-      var myRegex = /(\d\d\/\d\d\/201\d) - \d\d\/\d\d\/201\d/; //Find the first date in the comment string
-      if(!latest){
-        // Reverse order to get first comment but create a shallow copy to not break integration
-        commentList = commentList.slice(0).reverse();
-      }
-      var correctComment = commentList.find(function(comment){
-          var match = myRegex.exec(comment.data.text);
-          return match ? true : false;
-      });
-      if(correctComment){
-        var commentDateMatch = myRegex.exec(correctComment.data.text);
-        var commentDate = commentDateMatch[1];
-        var commMoment = moment(commentDate, "MM/DD/YYYY").toISOString();
-        return commMoment;
-      } else {
-        return false;
-      }
-    }
-
-    findPrevMoveDateFromComments(opts){
-      var actionList = opts.actionList;
-      var cardCreationDate = opts.cardCreationDate;
-      var commentList = opts.commentList;
-      var correctComment = false;
-      if(commentList){
-        var correctComment = this.checkCommentsForDates(commentList, true);
-      }
-      if(correctComment){
-        return correctComment;
-      } else if(actionList) {
-        if(actionList.length > 1){
-        return actionList[0].date;
-        }
-      } else {
-        if(opts.cardCreationDate){
-          return cardCreationDate;
-        } else {
-          return moment("01/01/2016", "MM/DD/YYYY").toISOString();
-        }
-      }
-    }
-
     inFinalList(listID){
       var deferred = Q.defer();
       var self = this;
@@ -174,29 +115,8 @@ class CardRecorder extends MyTrello {
       return deferred.promise;
     }
 
-    hasMovedCheck(actionList) {
-        var updated = false;
-        var moves = actionList.filter(function(a) {
-            return 'listBefore' in a.data;
-        });
-        if (moves.length) updated = moves;
-        return updated;
-    }
-
     getPreviousList(cardAction) {
         return cardAction.data.listBefore.name;
-    }
-
-    calcTotalDays(commentList, nowMoment){
-        var firstDate = this.checkCommentsForDates(commentList, false);
-        if(firstDate){
-          var dayDif = this.calculateDateDifference(10, firstDate, nowMoment)
-          //expected not actually needed, in the future could say total days expected
-          return dayDif[1];
-        } else {
-          return 0;
-        }
-
     }
 
     decideCommentType(card, finalList, daysSinceUpdate, hasMoved, prevMove, updateActions, totalDays, nowMoment){
@@ -241,7 +161,7 @@ class CardRecorder extends MyTrello {
           return stage.name == dateList;
         });
         var expectedTime = stage.expected_time;
-        var diffArray = this.calculateDateDifference(expectedTime, fromDate, toDate);
+        var diffArray = DCH.calculateDateDifference(expectedTime, fromDate, toDate);
         var differenceFromExpected = diffArray[0];
         var timeTaken = diffArray[1];
         var comment = this.buildComment(differenceFromExpected, expectedTime, fromDate, toDate, nameList, timeTaken, totDays);
@@ -251,24 +171,6 @@ class CardRecorder extends MyTrello {
             deferred.resolve(comment);
         }
         return deferred.promise;
-    }
-
-    findHolidaysBetweenDates(fromDate, toDate){
-      var count = 0;
-      holidays.forEach(function(holiday){
-        if(moment(holiday.date.toISOString(), ["YYYY-M-D", "YYYY-MM-DD", "YYYY-MM-D", "YYYY-M-DD"]).isBetween(fromDate, toDate, 'day')){
-          count++;
-        }
-      });
-      return count;
-    }
-
-    calculateDateDifference(expected, prevMove, recentMove) {
-        var fromDate = new Date(prevMove);
-        var toDate = new Date(recentMove);
-        var diffDays = instadate.differenceInWorkDays(fromDate, toDate);
-        var diffDays = diffDays - this.findHolidaysBetweenDates(fromDate, toDate);
-        return [diffDays - expected, diffDays];
     }
 
     buildComment(dateDiff, expected, prevMove, recentMove, lastList, actual, totalDays) {
@@ -290,6 +192,5 @@ class CardRecorder extends MyTrello {
         return deferred.promise;
     }
 }
-
 
 module.exports = CardRecorder;
