+++ app/index.js
@@ -3,3 +3,4 @@ require("dotenv").config();
 exports.StageManager = require("./stagemanager.js");
 exports.CardRecorder = require("./cardrecord.js");
 exports.CardCreator = require("./cardcreator.js");
+exports.DateCommentHelpers = require("./date-comment-helpers.js");
